+++ prometheus-manager.sh
@@ -4,41 +4,57 @@ MODT="Welcome to Prometheus manager v1.0.0!"
 # Set default values
 SYSTEM_ARCH=amd64 # -> can be changed by script argument -a arm64
 
-# Retrieve versions from .env file in form of:
+LOG_LEVEL=0
+
+# Retrieve versions from .versions file in form of:
 # NODE_EXPORTER_VERSION=1.3.1
 # PROMETHEUS_VERSION=2.33.5
-file=$(cat .env)
+file=$(cat .versions)
 for line in $file
 do
     eval "${line%=*}"="${line##*=}"
 done
 
-NODE_EXPORTER_PORT=9500
-PROMETHEUS_PORT=9590
+DISPLAY_VERSIONS=false
+DISPLAY_NODE_EXPORTER=false
+DISPLAY_BLACKBOX_EXPORTER=false
+DISPLAY_PROMETHEUS=false
+DISPLAY_ALERTMANAGER=false
+
+UPDATE_VERSIONS=false
+UPDATE_NODE_EXPORTER=false
+UPDATE_BLACKBOX_EXPORTER=false
+UPDATE_PROMETHEUS=false
+UPDATE_ALERTMANAGER=false
 
 INSTALL=false
 
 EXECUTE=false
+
+NODE_EXPORTER_PORT=9500
+PROMETHEUS_PORT=9590
 
 NODE_TRIGGER=false # -> can be changed either by script argument -n or -N 1.1.2 or -b
 PROMETHEUS_TRIGGER=false # -> can be changed either by script argument -p or -P 2.27.1 or -b
 
 function usage {
         echo "Usage: $(basename "$0") [<flags>]" 2>&1
-        echo '   -h, --help                 Show this help context'
-        echo '   -V, --versions             Retrieve versions of all available apps (prometheus, node_exporter, etc.)'
-        echo '   -s, --status               Prompt services status'
-        echo '   --ports                    List all ports actually used'
-        echo '   --arch arm64               Set architecture, default is amd64'
-        echo '   -I, --install              Install services combined with corresponding params e.g. --all'
-        echo '   -E, --exec                 Execute services combined with corresponding params e.g. --all'
-        echo '   --all                      Process script for all available apps (prometheus, node_exporter, etc.)'
-        echo '   -n, --node                 Process for node_exporter'
-        echo '   -N [<version>]             Specify node_exporter version'
-        echo '   -p, --prom                 Process for Prometheus'
-        echo '   -P [<version>]             Specify prometheus version'
-        echo '   -k, --kill                 Stop daemons for both prometheus and node_exporter'
-        echo '   --remove-all               Remove all data, users and services'
+        echo '   -h, --help                          Show this help context'
+        echo '   -v, --verbose                       Adaptative verbose mode (-vv for WARN, -vvv for INFO, -vvvv for full debugging)'
+        echo '   -V, --versions [all|<app>]          Display versions of all available apps (prometheus, node_exporter, etc.)'
+        echo '   -U, --update-versions [all|<app>]   Retrieve and override versions numbers for selected apps (prometheus, node_exporter, etc.)'
+        echo '   -s, --status                        Prompt services status'
+        echo '   --ports                             List all ports actually used'
+        echo '   --arch arm64                        Set architecture, default is amd64'
+        echo '   -I, --install                       Install services combined with corresponding params e.g. --all'
+        echo '   -E, --exec                          Execute services combined with corresponding params e.g. --all'
+        echo '   --all                               Process script for all available apps (prometheus, node_exporter, etc.)'
+        echo '   -n, --node                          Process for node_exporter'
+        echo '   -N [<version>]                      Specify node_exporter version'
+        echo '   -p, --prom                          Process for Prometheus'
+        echo '   -P [<version>]                      Specify prometheus version'
+        echo '   -k, --kill                          Stop daemons for both prometheus and node_exporter'
+        echo '   --remove-all                        Remove all data, users and services'
         exit 1
 }
 
@@ -50,9 +66,81 @@ function flags() {
         usage
         exit 1
         ;;
-      -V|--versions)
-        get_versions
+      -v|--verbose)
+        LOG_LEVEL=1
         shift # argument
+        ;;
+      -vv)
+        LOG_LEVEL=2
+        shift # argument
+        ;;
+      -vvv)
+        LOG_LEVEL=3
+        shift # argument
+        ;;
+      -vvvv)
+        LOG_LEVEL=4
+        shift # argument
+        ;;
+      -V|--versions)
+        check_options_mandatory "$2"
+        DISPLAY_VERSIONS_SELECTION="$2"
+        DISPLAY_VERSIONS=true
+        shift # argument
+        shift # value
+        if [ "$DISPLAY_VERSIONS_SELECTION" == "all" ]
+        then
+          DISPLAY_NODE_EXPORTER=true
+          DISPLAY_BLACKBOX_EXPORTER=true
+          DISPLAY_PROMETHEUS=true
+          DISPLAY_ALERTMANAGER=true
+        fi
+        if [ "$DISPLAY_VERSIONS_SELECTION" == "node" ]
+        then
+          DISPLAY_NODE_EXPORTER=true
+        fi
+        if [ "$DISPLAY_VERSIONS_SELECTION" == "blackbox" ]
+        then
+          DISPLAY_BLACKBOX_EXPORTER=true
+        fi
+        if [ "$DISPLAY_VERSIONS_SELECTION" == "prom" ]
+        then
+          DISPLAY_PROMETHEUS=true
+        fi
+        if [ "$DISPLAY_VERSIONS_SELECTION" == "alert" ]
+        then
+          DISPLAY_ALERTMANAGER=true
+        fi
+        ;;
+      -U|--update-versions)
+        check_options_mandatory "$2"
+        UPDATE_VERSIONS_SELECTION="$2"
+        UPDATE_VERSIONS=true
+        shift # argument
+        shift # value
+        if [ "$UPDATE_VERSIONS_SELECTION" == "all" ]
+        then
+          UPDATE_NODE_EXPORTER=true
+          UPDATE_BLACKBOX_EXPORTER=true
+          UPDATE_PROMETHEUS=true
+          UPDATE_ALERTMANAGER=true
+        fi
+        if [ "$UPDATE_VERSIONS_SELECTION" == "node" ]
+        then
+          UPDATE_NODE_EXPORTER=true
+        fi
+        if [ "$UPDATE_VERSIONS_SELECTION" == "blackbox" ]
+        then
+          UPDATE_BLACKBOX_EXPORTER=true
+        fi
+        if [ "$UPDATE_VERSIONS_SELECTION" == "prom" ]
+        then
+          UPDATE_PROMETHEUS=true
+        fi
+        if [ "$UPDATE_VERSIONS_SELECTION" == "alert" ]
+        then
+          UPDATE_ALERTMANAGER=true
+        fi
         ;;
       -s|--status)
         get_status
@@ -63,6 +151,7 @@ function flags() {
         shift # argument
         ;;
       --arch)
+        check_options_mandatory "$2"
         SYSTEM_ARCH="$2"
         shift # argument
         shift # value
@@ -85,6 +174,7 @@ function flags() {
         shift # argument
         ;;
       -N)
+        check_options_mandatory "$2"
         NODE_EXPORTER_VERSION="$2"
         NODE_TRIGGER=true
         shift # argument
@@ -95,6 +185,7 @@ function flags() {
         shift # argument
         ;;
       -P)
+        check_options_mandatory "$2"
         PROMETHEUS_VERSION="$2"
         PROMETHEUS_TRIGGER=true
         shift # argument
@@ -163,63 +254,241 @@ function flags() {
 }
 
 
+function check_options_mandatory() {
+    if [ "${1::1}" == "-" ]
+    then
+      echo "Value is mandatory for argument before $1"
+      exit 1
+    fi
+
+    if [ -z "$1" ]
+    then
+      echo "Value is mandatory for final argument"
+      exit 1
+    fi
+}
+
+
 function check_root_rights(){
   if [ `id -u` -ne 0 ]
   then
-    echo "[WARN] This script require sudo rights to run smoothly"
+    echo "[ERROR] This script require sudo rights to run smoothly"
   fi
 }
 
 
-function get_node_version() {
-  NODE_EXPORTER_VERSION_CURLED=$(curl -s https://api.github.com/repos/prometheus/node_exporter/releases/latest |
-  grep "tag_name" | awk '{printf substr($2, 3, length($2)-4)}')
+function retrieve_node_version() {
+    if [ $LOG_LEVEL -gt 3 ]
+    then
+      echo '[DEBUG] Retrieving node_exporter version...'
+    fi
+    NODE_EXPORTER_VERSION_CURLED="$(curl -s https://api.github.com/repos/prometheus/node_exporter/releases/latest |
+    grep 'tag_name' | awk '{printf substr($2, 3, length($2)-4)}')"
 
-  if [[ -n $NODE_EXPORTER_VERSION_CURLED ]]
+    if [[ -z $NODE_EXPORTER_VERSION_CURLED ]]
+    then
+      echo '[ERROR] System was not able to retrieve node_exporter version check your internet connection'
+      exit 1
+    fi
+
+    if [ $LOG_LEVEL -gt 2 ]
+    then
+      echo "[INFO] node_exporter version retrieved: " "$NODE_EXPORTER_VERSION_CURLED"
+    fi
+}
+
+
+function retrieve_blackbox_version() {
+    if [ $LOG_LEVEL -gt 3 ]
+    then
+      echo '[DEBUG] Retrieving blackbox_exporter version...'
+    fi
+    BLACKBOX_EXPORTER_VERSION_CURLED="$(curl -s https://api.github.com/repos/prometheus/blackbox_exporter/releases/latest |
+    grep 'tag_name' | awk '{printf substr($2, 3, length($2)-4)}')"
+
+    if [[ -z $BLACKBOX_EXPORTER_VERSION_CURLED ]]
+    then
+      echo '[ERROR] System was not able to retrieve blackbox_exporter version check your internet connection'
+      exit 1
+    fi
+
+    if [ $LOG_LEVEL -gt 2 ]
+    then
+      echo "[INFO] blackbox_exporter version retrieved: " "$BLACKBOX_EXPORTER_VERSION_CURLED"
+    fi
+}
+
+
+function retrieve_prometheus_version() {
+    if [ $LOG_LEVEL -gt 3 ]
+    then
+      echo '[DEBUG] Retrieving prometheus version...'
+    fi
+    PROMETHEUS_VERSION_CURLED="$(curl -s https://api.github.com/repos/prometheus/prometheus/releases/latest |
+    grep 'tag_name' | awk '{printf substr($2, 3, length($2)-4)}')"
+
+    if [[ -z $PROMETHEUS_VERSION_CURLED ]]
+    then
+      echo '[ERROR] System was not able to retrieve prometheus version check your internet connection'
+      exit 1
+    fi
+
+    if [ $LOG_LEVEL -gt 2 ]
+    then
+      echo "[INFO] prometheus version retrieved: " "$PROMETHEUS_VERSION_CURLED"
+    fi
+}
+
+
+function retrieve_alertmanager_version() {
+    if [ $LOG_LEVEL -gt 3 ]
+    then
+      echo '[DEBUG] Retrieving alertmanager version...'
+    fi
+    ALERTMANAGER_VERSION_CURLED="$(curl -s https://api.github.com/repos/prometheus/alertmanager/releases/latest |
+    grep 'tag_name' | awk '{printf substr($2, 3, length($2)-4)}')"
+
+    if [[ -z $ALERTMANAGER_VERSION_CURLED ]]
+    then
+      echo '[ERROR] System was not able to retrieve alertmanager version check your internet connection'
+      exit 1
+    fi
+
+    if [ $LOG_LEVEL -gt 2 ]
+    then
+      echo "[INFO] alertmanager version retrieved: " "$ALERTMANAGER_VERSION_CURLED"
+    fi
+}
+
+
+function ensure_versions() {
+  # Test if file is correctly filled or if update is request
+  if [ "${#NODE_EXPORTER_VERSION}" -lt 5 ] || $UPDATE_NODE_EXPORTER
   then
+    retrieve_node_version
     NODE_EXPORTER_VERSION=$NODE_EXPORTER_VERSION_CURLED
   fi
 
+  if [ "${#BLACKBOX_EXPORTER_VERSION}" -lt 5 ] || $UPDATE_BLACKBOX_EXPORTER
+  then
+    retrieve_blackbox_version
+    BLACKBOX_EXPORTER_VERSION=$BLACKBOX_EXPORTER_VERSION_CURLED
+  fi
+
+  if [ "${#PROMETHEUS_VERSION}" -lt 5 ] || $UPDATE_PROMETHEUS
+  then
+    retrieve_prometheus_version
+    PROMETHEUS_VERSION=$PROMETHEUS_VERSION_CURLED
+  fi
+
+  if [ "${#ALERTMANAGER_VERSION}" -lt 5 ] || $UPDATE_ALERTMANAGER
+  then
+    retrieve_alertmanager_version
+    ALERTMANAGER_VERSION=$ALERTMANAGER_VERSION_CURLED
+  fi
+}
+
+
+function store_actual_versions() {
+  if [ $LOG_LEVEL -gt 3 ]
+  then
+    echo '[DEBUG] Storing new versions in .versions file...'
+  fi
+  cat > .versions << EOM
+NODE_EXPORTER_VERSION=$NODE_EXPORTER_VERSION
+BLACKBOX_EXPORTER_VERSION=$BLACKBOX_EXPORTER_VERSION
+PROMETHEUS_VERSION=$PROMETHEUS_VERSION
+ALERTMANAGER_VERSION=$ALERTMANAGER_VERSION
+EOM
+
+  chmod 666 .versions
+}
+
+
+function update_versions() {
+  ensure_versions
+  store_actual_versions
+}
+
+
+function display_node_versions() {
+  if [ -z "$NODE_EXPORTER_VERSION_CURLED" ]
+  then
+    NODE_EXPORTER_VERSION_CURLED="Run with -U, --update-versions command argument"
+  fi
   printf "Highest available version for node_exporter is: %s\n" "$NODE_EXPORTER_VERSION_CURLED"
 
   printf "Installation version for node_exporter will be: %s\n" "$NODE_EXPORTER_VERSION"
 
-  printf "Actual version of node_exporter is:\n"
-  /usr/local/bin/node_exporter --version
+  INSTALLED_NODE_EXPORTER_VERSION=$(/usr/local/bin/node_exporter --version 2>&1 | awk 'NR==1 {printf $3}')
+  if [ "${#INSTALLED_NODE_EXPORTER_VERSION}" -lt 5 ]
+  then
+    INSTALLED_NODE_EXPORTER_VERSION="Not installed"
+  fi
+  printf "Actual version of node_exporter is: %s\n\n" "$INSTALLED_NODE_EXPORTER_VERSION"
 }
 
 
-function get_prometheus_version() {
-  PROMETHEUS_VERSION_CURLED=$(curl -s https://api.github.com/repos/prometheus/prometheus/releases/latest |
-  grep "tag_name" | awk '{printf substr($2, 3, length($2)-4)}')
-
-  if [[ -n $PROMETHEUS_VERSION_CURLED ]]
+function display_blackbox_versions() {
+  if [ -z "$BLACKBOX_EXPORTER_VERSION_CURLED" ]
   then
-    PROMETHEUS_VERSION=$PROMETHEUS_VERSION_CURLED
+    BLACKBOX_EXPORTER_VERSION_CURLED="Run with -U, --update-versions command argument"
   fi
+  printf "Highest available version for blackbox_exporter is: %s\n" "$BLACKBOX_EXPORTER_VERSION_CURLED"
 
-  printf "\nHighest available version for Prometheus is: %s\n" "$PROMETHEUS_VERSION_CURLED"
+  printf "Installation version for blackbox_exporter will be: %s\n" "$BLACKBOX_EXPORTER_VERSION"
+
+  INSTALLED_BLACKBOX_EXPORTER_VERSION=$(/usr/local/bin/blackbox_exporter --version 2>&1 | awk 'NR==1 {printf $3}')
+  if [ "${#INSTALLED_BLACKBOX_EXPORTER_VERSION}" -lt 5 ]
+  then
+    INSTALLED_BLACKBOX_EXPORTER_VERSION="Not installed"
+  fi
+  printf "Actual version of blackbox_exporter is: %s\n\n" "$INSTALLED_BLACKBOX_EXPORTER_VERSION"
+}
+
+
+function display_prometheus_versions() {
+  if [ -z "$PROMETHEUS_VERSION_CURLED" ]
+  then
+    PROMETHEUS_VERSION_CURLED="Run with -U, --update-versions command argument"
+  fi
+  printf "Highest available version for Prometheus is: %s\n" "$PROMETHEUS_VERSION_CURLED"
 
   printf "Installation version for Prometheus will be: %s\n" "$PROMETHEUS_VERSION"
 
-  /usr/local/bin/prometheus --version | awk 'NR==1 {printf "Actual version of Prometheus is: %s\n", $3}'
-  test -f /usr/local/bin/promtool && echo Promtool is present
-  echo
+  INSTALLED_PROMETHEUS_VERSION=$(/usr/local/bin/prometheus --version 2>&1 | awk 'NR==1 {printf $3}')
+  if [ "${#INSTALLED_PROMETHEUS_VERSION}" -lt 5 ]
+  then
+    INSTALLED_PROMETHEUS_VERSION="Not installed"
+  fi
+  printf "Actual version of prometheus is: %s\n\n" "$INSTALLED_PROMETHEUS_VERSION"
+  test -f /usr/local/bin/promtool && printf "Promtool is present\n"
 }
 
 
-function set_versions() {
-  cat > .env << EOM
-NODE_EXPORTER_VERSION=$NODE_EXPORTER_VERSION
-PROMETHEUS_VERSION=$PROMETHEUS_VERSION
-EOM
+function display_alertmanager_versions() {
+  if [ -z "$ALERTMANAGER_VERSION_CURLED" ]
+  then
+    ALERTMANAGER_VERSION_CURLED="Run with -U, --update-versions command argument"
+  fi
+  printf "Highest available version for alertmanager is: %s\n" "$ALERTMANAGER_VERSION_CURLED"
+
+  printf "Installation version for alertmanager will be: %s\n" "$ALERTMANAGER_VERSION"
+
+  INSTALLED_ALERTMANAGER_VERSION=$(/usr/local/bin/alertmanager --version 2>&1 | awk 'NR==1 {printf $3}')
+  if [ "${#INSTALLED_ALERTMANAGER_VERSION}" -lt 5 ]
+  then
+    INSTALLED_ALERTMANAGER_VERSION="Not installed"
+  fi
+  printf "Actual version of alertmanager is: %s\n\n" "$INSTALLED_ALERTMANAGER_VERSION"
 }
 
 
-function get_versions() {
-  get_node_version
-  get_prometheus_version
-  set_versions
+function display_versions() {
+  if $DISPLAY_NODE_EXPORTER; then display_node_versions; fi
+  if $DISPLAY_BLACKBOX_EXPORTER; then display_blackbox_versions; fi
+  if $DISPLAY_PROMETHEUS; then display_prometheus_versions; fi
+  if $DISPLAY_ALERTMANAGER; then display_alertmanager_versions; fi
 }
 
 
@@ -433,6 +702,14 @@ EOM
 function main() {
   printf 'Selected architecture is %s\n\n' $SYSTEM_ARCH
 
+  if $UPDATE_VERSIONS; then
+    update_versions
+  fi
+
+  if $DISPLAY_VERSIONS; then
+    display_versions
+  fi
+
   if $INSTALL && $NODE_TRIGGER; then
     update_node_exporter
   fi
@@ -448,7 +725,7 @@ function main() {
   fi
 
   if $EXECUTE; then
-    get_versions
+    display_versions
   fi
 }